+++ src/main/java/game/Application.java
@@ -1,5 +1,9 @@
 package game;
 
+import game.ui.RacingConsole;
+import game.ui.RacingInputImpl;
+import game.ui.RacingOutputImpl;
+
 public class Application {
     public static void main(String[] args) {
         RacingConsole racingConsole = new RacingConsole(new RacingInputImpl(), new RacingOutputImpl());
