+++ src/main/java/game/domain/Cars.java
@@ -1,4 +1,4 @@
-package game;
+package game.domain;
 
 import java.util.Arrays;
 import java.util.List;
