+++ src/main/java/game/domain/GameResult.java
@@ -1,4 +1,4 @@
-package game;
+package game.domain;
 
 import java.util.ArrayList;
 import java.util.List;
