+++ src/main/java/game/domain/RoundResult.java
@@ -1,4 +1,6 @@
-package game;
+package game.domain;
+
+import game.domain.DriveResult;
 
 import java.util.ArrayList;
 import java.util.Comparator;
