+++ src/main/java/game/ui/RacingInputImpl.java
@@ -1,4 +1,4 @@
-package game;
+package game.ui;
 
 import java.io.BufferedReader;
 import java.io.IOException;
