+++ src/main/java/game/ui/RacingOutput.java
@@ -1,4 +1,6 @@
-package game;
+package game.ui;
+
+import game.domain.GameResult;
 
 public interface RacingOutput {
     void printRequestCarNames();
