+++ src/test/java/game/CarTests.java
@@ -1,5 +1,7 @@
 package game;
 
+import game.domain.Car;
+import game.domain.DriveResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
